+++ google-modules/soc/gs/drivers/dma-buf/heaps/samsung/cma_heap.c
@@ -27,7 +27,7 @@ struct cma_heap {
 };
 
 static struct dma_buf *cma_heap_allocate(struct dma_heap *heap, unsigned long len,
-					 unsigned long fd_flags, unsigned long heap_flags)
+					 u32 fd_flags, u64 heap_flags)
 {
 	struct samsung_dma_heap *samsung_dma_heap = dma_heap_get_drvdata(heap);
 	struct cma_heap *cma_heap = samsung_dma_heap->priv;
